Fix bug causing removed books to be readded
Removing multiple books at random and then adding another book caused
previously removed books to be readded to the page. This was due to
the indexes on the HTML elements falling out of sync with the
myLibrary array. Calling the updateBookContainer function each time a
book is removed corrects the issue.